Fix jsonify vaulted strings
